renamed i10n.js to languageStrings.js and updated imports accordingly.
